code: double

+++ main.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int main()
 {
     cout << "Enter A and B: ";
-    int a, b, max;
+    double a, b, max;
     cin >> a >> b;
     cout << "A + B = " << a + b << '\n'
      << "A - B = " << a - b << '\n'
